# PERF: leverage tzlocal package to provide 2000x speedup for dateutil.tz.tzlocal operations