--- a/pandas/tests/tslibs/test_timezones.py
+++ b/pandas/tests/tslibs/test_timezones.py
@@ -49,6 +49,27 @@ def test_tzlocal_offset():
     assert ts.value + offset == Timestamp("2011-01-01").value
 
 
+def test_tzlocal_package():
+    orig_tzlocal_tz = timezones.get_tzlocal_tz(None)
+    # Check that we fall back to non-tzlocal code path if user has an invalid
+    # system tz set
+    timezones._set_tzlocal_tz('foo/bar')
+    ts = Timestamp("2011-01-01", tz=dateutil.tz.tzlocal())
+    assert ts.tz == dateutil.tz.tzlocal()
+
+    utc = pytz.utc
+    for tz in pytz.all_timezones:
+        # Check that all pytz timezones work
+        ts_tz = Timestamp("2011-01-01", tz=tz)
+        timezones._set_tzlocal_tz(tz)
+        ts_local = Timestamp("2011-01-01", tz=dateutil.tz.tzlocal())
+
+        assert ts_tz.astimezone(utc) == ts_local.astimezone(utc)
+        assert ts_local.tz == dateutil.tz.tzlocal()
+
+    timezones._set_tzlocal_tz(orig_tzlocal_tz.zone)
+
+
 @pytest.fixture(params=[
     (pytz.timezone("US/Eastern"), lambda tz, x: tz.localize(x)),
     (dateutil.tz.gettz("US/Eastern"), lambda tz, x: x.replace(tzinfo=tz))