… pandas/_libs/src/tzlocal/windows_tz.py