This is a follow-up to https://github.com/pandas-dev/pandas/pull/24491, where it was noted that `tzlocal` operations are notably slower:
```
[ 53.12%] ··· timeseries.DatetimeIndex.time_add_timedelta                                                                                                                 ok
[ 53.12%] ··· ============ =============
               index_type
              ------------ -------------
                  dst         471±8μs
                repeated    3.52±0.05ms
                tz_aware    3.78±0.07ms
              **tz_local     5.77±0.01s**
                tz_naive    1.98±0.03ms
              ============ =============
```
This PR makes them indistinguishable from other timezones:
```
[ 60.00%] ··· timeseries.DatetimeIndex.time_add_timedelta                                                                                                                 ok
[ 60.00%] ··· ============ =============
               index_type
              ------------ -------------
                  dst         465±8μs
                repeated    3.55±0.07ms
                tz_aware    3.65±0.05ms
                tz_local    3.68±0.04ms
                tz_naive    1.81±0.05ms
              ============ =============
```
To do this, we add a new optional dependency: `tzlocal`. This package queries the OS to try and identify a `pytz` compatible timezone name for the local timezone. If this is successful, we can use the `pytz` timezone instead of `dateutil.tz.tzlocal()`, yielding a huge speedup as the latter invokes `datetime.datetime()`. In cases where the local timezone is also UTC, further speedups are possible.

In all cases, the `tzlocal()` object is maintained as the `.tz` to preserve any semantic meaning.

One unresolved issue is an incompatibility with `python-dateutil` versions prior to `2.6.0`. I'm unable to reproduce the error locally, but can repeatably get this error on Travis:
```
    def test_timetz_accessor(self, tz_naive_fixture):
        # GH21358
        tz = timezones.maybe_get_tz(tz_naive_fixture)
    
        expected = np.array([time(10, 20, 30, tzinfo=tz), pd.NaT])
    
        index = DatetimeIndex(['2018-06-04 10:20:30', pd.NaT], tz=tz)
        result = index.timetz
    
>       tm.assert_numpy_array_equal(result, expected)
pandas/tests/indexes/datetimes/test_timezones.py:824: 
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
pandas/core/dtypes/missing.py:404: in array_equivalent
    ensure_object(left.ravel()), ensure_object(right.ravel()))
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ 
>   if not (PyObject_RichCompareBool(x, y, Py_EQ) or
E   TypeError: can't compare offset-naive and offset-aware times
```
While this only shows up on the 2.7 build, I believe this is due to it being the only one actually building against `python-dateutil=2.5.0`. Any version `>=2.6.0` does not exhibit this error.

Finally, here is the `asv` benchmark against `v0.24.0rc1`:
```
$ asv compare v0.24.0rc1 HEAD -s --sort ratio --only-changed
       before           after         ratio
     [fdc4db25]       [56e5cc0d]
     <v0.24.0rc1^0>       <tzlocal>
-        68.4±5μs         17.9±1μs     0.26  timeseries.DatetimeIndex.time_get('tz_local')
-      5.54±0.02s          188±2ms     0.03  timeseries.DatetimeIndex.time_to_pydatetime('tz_local')
-      7.13±0.02s       30.2±0.7ms     0.00  timeseries.DatetimeIndex.time_normalize('tz_local')
-      6.00±0.05s       23.7±0.3ms     0.00  timeseries.DatetimeIndex.time_to_time('tz_local')
-      6.00±0.01s       21.3±0.3ms     0.00  timeseries.DatetimeIndex.time_to_date('tz_local')
-      5.98±0.06s      3.68±0.04ms     0.00  timeseries.DatetimeIndex.time_add_timedelta('tz_local') 
-      5.96±0.05s      2.49±0.05ms     0.00  timeseries.DatetimeIndex.time_timeseries_is_month_start('tz_local')
```

- [ ] closes #xxxx
- [ ] tests added / passed
- [ ] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry