@qwhelan 2019-01-12
@mroeschke Agreed, it's worth mentioning that this mostly fixes the top performance regression between `v0.23.4` and `v0.24.0rc1`:
```
       before           after         ratio
     [04095216]       [fdc4db25]
     <v0.23.4^0>       <v0.24.0rc1^0>
+     3.91±0.01ms       18.8±0.01s  4819.04  timeseries.DatetimeIndex.time_add_timedelta('tz_local')
```
It'd be relatively simple to vendor it as it's only a handful of files. It's MIT licensed, so should be fine but worth mentioning now.

I haven't been able to reproduce the `python-dateutil` conflict locally thus far, but I'll dig into that further to see if there's an easy workaround.