@jreback 2019-01-12
if we bump to 2.6.0 would u still need tzlocal?