@qwhelan 2019-01-12
@jreback Yes, there's just a hard-to-reproduce conflict with 2.5.0-2.5.3. Not sure how serious it is at this point.

@jbrockmendel Regarding accuracy, it's pulling from system files like `/etc/timezone`, `/var/db/zoneinfo`, etc so should accurately reflect the system timezone. However, given that it's checking so many places it's entirely possible a user has different timezones configured in different files.

There's no distinction between `America/Los_Angeles` and `US/Pacific` other than the latter being deprecated (https://en.wikipedia.org/wiki/List_of_tz_database_time_zones).