@mroeschke 2019-01-13
Although tzlocal hurts portability (unless that's the intension), I think pandas should ensure that timezones in == timezones out and shouldn't make any assumptions or transformation when returning the timezone (i.e tzlocal in, tzlocal out; US/Pacific in, US/Pacific out), which I can see your PR does @qwhelan .

Overall agreed that I'd be great to address this regression. Shot in the dark: I think the proper way to compare dateutil tzs is with `is` and not `==`, so maybe that's why you're seeing that error?