@pep8speaks 2019-01-30
Hello @qwhelan! Thanks for updating the PR.



 - In the file [`pandas/tests/series/test_datetime_values.py`](https://github.com/pandas-dev/pandas/blob/79bb80e20b2a824f71c36c274f4abc56eb6fdb36/pandas/tests/series/test_datetime_values.py), following are the PEP8 issues :

> [Line 541:9](https://github.com/pandas-dev/pandas/blob/79bb80e20b2a824f71c36c274f4abc56eb6fdb36/pandas/tests/series/test_datetime_values.py#L541): [E265](https://duckduckgo.com/?q=pep8%20E265) block comment should start with '# '