@qwhelan 2019-02-05
@jreback I'm finally able to reproduce `dateutil` 2.5.0 issue locally and think I have it resolved without needing to bump the dependency. Where would be the appropriate directory to vendor `tzlocal`?