@jreback 2019-02-05
put in pandas/tseries/_tzlocal.py i think - see how that works