@mroeschke 2019-02-05
@jreback How about `pandas/_libs/tslibs/*`? Could give way to cythonize some things?