@jbrockmendel 2019-02-05
> Could give way to cythonize some things?

Is there anything in particular that bears cythonizing?