@mroeschke 2019-02-05
Haven't looked closely, but I imagine most all functions here can be `cdef`'d given that it just interacts in `timezones.pyx`?