@WillAyd 2019-02-27
@qwhelan can you merge master and address build failures?