@jreback 2019-06-08
closing as stale. nice idea though, pls ping if you can continue.